requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1.0
python-dotenv>=1.0
dj-database-url >=0.5.0
djangorestframework>=3
djangorestframework_simplejwt>=5
pydantic>=2.0
python-decouple>=3.8
psycopg2-binary>=2.9
django-cors-headers>=4.3.0
pydantic[email]